requirements.txt
numpy
scipy
matplotlib
h5py
pandas
Pillow
